Factorial without recursion, BubbleSort, Factorial by recursion, 2Darrays, average,add, arithematic series

+++ add.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main() //main function
 {
-  int a, b;
+  float a, b;
 
   printf("Enter first number:\n");
-  scanf("%d", &a);
+  scanf("%f", &a);
   printf("Enter second number:\n");
-  scanf("%d", &b);
+  scanf("%f", &b);
 
-  printf("Addition of two numbers = %d", a + b);
+  printf("Addition of two numbers = %.2f", (a + b));
   return 0;
 }
